Changed the type of all IPv4 addresses from uint32_t to ipv4_addr_t.
This really exists only to ease the implementation of IPv6 by highlighting
all the instances where the addresses are.

+++ include/rail-flush.h
@@ -21,8 +21,8 @@
 
 # include <stdint.h>
 
-uint32_t
-  rail_flushrail(uint32_t len,
+size_t
+  rail_flushrail(size_t len,
                  int rail);
 
 #endif /* NBWORKS_RAILFLUSH_H */
